tree-construction/pending-spec-changes-plain-text-unsafe.dat

#data
<html>�<frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>

#data
<html> � <frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>

#data
<html>a�a<frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <body>
|     "a�a"

#data
<html>��<frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>

#data
<html>�
 <frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>
